+++ internal/pkg/sdk3rd/ctyun/cdn/api_query_cert_detail.go
@@ -7,7 +7,7 @@ import (
 )
 
 type QueryCertDetailRequest struct {
-	Id        *int32  `json:"id,omitempty"`
+	Id        *int64  `json:"id,omitempty"`
 	Name      *string `json:"name,omitempty"`
 	UsageMode *int32  `json:"usage_mode,omitempty"`
 }
